requirements.txt
streamlit
pandas
boto3
python-dotenv
matplotlib